--- src/test/java/com/mageddo/utils/dagger/mockito/haystack/Haystack.java
@@ -1,4 +1,0 @@
-package com.mageddo.utils.dagger.mockito.haystack;
-
-public class Haystack {
-}
